Rename the `Cycle\ORM\Command\Branch` namespace to `Cycle\ORM\Command\Special`
