+++ tests/ORM/Fixtures/UserSnapshotMapper.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Cycle\ORM\Tests\Fixtures;
 
-use Cycle\ORM\Command\Branch\Sequence;
-use Cycle\ORM\Command\Branch\WrappedStoreCommand;
+use Cycle\ORM\Command\Special\Sequence;
+use Cycle\ORM\Command\Special\WrappedStoreCommand;
 use Cycle\ORM\Command\CommandInterface;
 use Cycle\ORM\Command\StoreCommandInterface;
 use Cycle\ORM\Heap\Node;
